Added a small notice where to find the registration code.
There is a procedure which is disabled ATM. I'm not sure
if this procedure always work. For the time being, I think
using a minimal approach (as indicated by the change in this
commit) is probably enough.
We can enable the procedure, if all the things about SUSEConnect
has been settled.


git-svn-id: https://svn.opensuse.org/svn/opensuse-doc/trunk/documents/sle/en@13173 8ed6235d-527f-0410-a63b-cf03ab4a5eb2