vt bp: Remove mention of XP/Server 2003 (and other small fixes)
Before I fucked up^W^Wchanged this section the last time, it used
the phrasing "Windows Server 20xx". Then I swooped in, read that the
Microsoft KBs were for XP/Server 2003 and changed our text to that
despite the fact that the procedure probably has not changed too much
since. So let's use the more generic "Windows Server" without a version
number here instead.

Also comment a URL that is not reachable anymore and add a remark next
to other URL.