vt doc: add a note about user and group ID requirements
In a cluster of VM hosts participating in migration, the UID of
qemu user and GID of qemu, kvm, and libvirt groups need to be the
same value. Add a note about the requirement in the "Migration
Requirements" section of the VT doc.

While at it, remove the bogus commment about migrating between Intel
and AMD hosts.

Signed-off-by: Jim Fehlig <jfehlig@suse.com>