Remove `/connect/` from SMT URL
To configure the installer to use a registration proxy, the FQDN of SMT/RMT is enough. There's no need to append `/connect/`, since the SUSEConnect library takes care of that on its own. In fact, SUSEConnect does not handle paths, and even if `fqdn/connect` is given as the regurl, it will only use `fqdn`, and ignore the path. Giving a path can cause a bug when de-registering a system (bsc#954451).