Entities for doc.suse.com: Use dependent entities, remove SBP
The SBP entity was inconsistent with the others (it included /html) and
was never used -> removed for the moment.

Use new "sle-ha-12", "sles-sap-12" short URLs: This ensures that the
URLs will work during the beta phase while the 12 SP5 docs are not up
yet as well.

Introduce new sles, sled, ses, suma URL entities.