vt_best_practices: reword the CPU assignment/overcommit section
The introduction of the "Assigning CPUs" section was just strongly
discouraging CPU overcommit, without much explanation of why this was
the case.

Add some hints about that, and also include some basic advice on how to
decide whether overcommitting may or not be sustainable.

Add the indication to absolutely avoid overcommitting by more than 8
times the number of CPUs. This, right now, comes from a mix of common
sense and experience, but it should be good to have tests that actually
check whether "8 times" is really the best limit.

Signed-off-by: Dario Faggioli <dfaggioli@suse.com>