vt doc: Improve section describing virtual controllers
The description of the xenbus controller and its maxGrantFrames attribute
is a bit confusing since it is under a section named "Managing Guest Memory
Allocation (Xen only)". It should be under a section discussing controller
configuration.

Move the section near other sections describing device configuration and
rename it to "Configuring Controller Devices". Reword the section a bit to
make it more general while retaining the Xen-specific information.

Signed-off-by: Jim Fehlig <jfehlig@suse.com>