Improvements for virsh chapter of virtualization guide (#585)
* vt doc: Improve section describing virtual controllers

The description of the xenbus controller and its maxGrantFrames attribute
is a bit confusing since it is under a section named "Managing Guest Memory
Allocation (Xen only)". It should be under a section discussing controller
configuration.

Move the section near other sections describing device configuration and
rename it to "Configuring Controller Devices". Reword the section a bit to
make it more general while retaining the Xen-specific information.

Signed-off-by: Jim Fehlig <jfehlig@suse.com>

* vt-doc: Add description for maxEventChannel attribute of xenbus controller

See bsc#1168767 for more details.

Signed-off-by: Jim Fehlig <jfehlig@suse.com>

* vt-doc: Add section describing hypervisor features

Add a section to Chapter 15 describing how to configure hypervisor
features with virsh.

See bsc#1167217 for more details.

Signed-off-by: Jim Fehlig <jfehlig@suse.com>