Partially revert 36a74fe4, we actually need MAIN files for many langs
The default English MAIN file unfortunately also includes an English
language code. That is not great, except in the sense that this greatly
confuses the Docserv2 automagic that finds out language codes and such.

ar-ar, cs-cz, hu-hu, pl-pl, and ru-ru only ship the Install Quick for
which we do not need the MAIN.SLEDS.xml currently when building. Hence,
we can continue to use symlinks for these languages.