GHA: Update workflow, everything is now in GitHub Actions, not Travis
Most of the behavior of the new workflow matches the behavior of the
old travis.sh script. Except of course, we now have builds running in
parallel.

Builds still only run for specified branches rather than for all
branches.

There's currently also the drawback that it's not possible to check
subdirs of a repo which impacts maintenance/ branches primarily.