Support to define the default skel removed from AutoYaST
The version of yast2-users committed to Tumbleweed by the end of June
2021, and that will be part of SLE-15-SP4, removed this attribute from
the AutoYaST profile for several reasons.

Basically, the management of skeleton directories is currently more
complex than it used to be. It does not depend only on the historical
SKEL setting at /etc/default/useradd. Although useradd still reads that
configuration, it does not offer any way to modify it since SLE-12 (when
useradd was moved from the dropped package "pwdutils" in favor of an
alternative implementation in the package "shadow"). So it was decided
that is not a good idea for AutoYaST to fiddle with that configuration.