vt-doc: Update screenshot of virt-manager CPUs dialog
The current dialog should reflect the default settings, even when
additional setting are expanded. E.g. if the "Copy host CPU configuration"
option is unchecked, the expanded model setting should show the default of
'host-model'. Also, a screenshot where the host machine only has a single
logical CPU doesn't provide a good example.

While at it, change the page name from 'Processor' to 'CPU'. The page is
identified as 'CPU' in virt-manager. It is fine to leave the doc section as
'Processor', but the page name should match virt-manager's naming.