virsh cpu and mem fixes (#936)
* vt-doc: Improve description of CPU configuration with virsh

The documentation around configuring VM CPUs with virsh only mentions
setting the number of CPUs. This commit add subsections describing
not only how to set the number of CPUs, but also how to set the CPU
model and features.

* vd-doc: Include an example of hotplugging CPUs

The virsh CPU configuration section describes how to change
the number of CPUs on an inactive VM. This commit describes how
to change the number of CPUs on a running VM and includes an
example procedure.

* vt-doc: Add note about CPU counts greater than 255

Setting the number of CPUs above 255 requires special VM
configuration. Add a note about it in the virsh CPU configuration
section and include an example.

* vt-doc: Improve description of memory configuration with virsh

The documentation around configuring memory with virsh only mentions
the <memory> setting. This change adds information about the
<currentMemory> setting and adds an example of changing VM memory
allocation at run-time with 'virsh setmem'.

* vt-doc: Add note to virsh section about large memory VMs

Similar to the virt-manager section, add a note to the virsh section
about the need for host-passthrough CPU model whem VM memory
allocation is >= 4TB.

* Apply suggestions from code review

Suggestions from tbazant's review of VT guide changes.

Co-authored-by: Tomáš Bažant <tbazant@suse.cz>

* Apply suggestions from code review

Co-authored-by: Tomáš Bažant <tbazant@suse.cz>