vt-doc: Fix virt-install example for Xen
The virt-install example for Xen (Example 8.2 in the Virtualization Guide)
does not work as is. PV is the default machine type for Xen, but PV guests
cannot boot from cdrom. HVM guests can boot from cdrom, so change the
example to create an HVM guest.

While at it, update the examples to install SLE15 SP2 instead of an old
distro like SLED12.