The section sec-util-processes is in a SLED book
This book is in the openSUSE documentation but it refers to a book that is not and its title does not tell what product it belongs to.  Add the correct product name to the reference (it is not a hyperlink) to give the reader a clue where to find it.