Update ay_partitioning.xml (#1094)
There is an unclosed section the example for multipath.
There is a conflicting note under Multipath support that mentions CT_DMMULTIPATH is deprecated, but the TID at https://www.suse.com/support/kb/doc/?id=000019015 says otherwise. Wouldn't it be safer to add CT_DMMULTIPATH in the list and mention that it's deprecated as well?