Change library names with the name of the packages.
Previously, the ULP docs referenced the livepatchable libraries by a
mix of the package name and the .so files. This is confusing, as
people reported libgcrypt being livepatchable when in fact it was not
because openssl provides libcrypt.so. Hence, update the name to always
mention the package name.

Fixes bsc#120636.

Signed-off-by: Giuliano Belinassi <gbelinassi@suse.de>