tuning_kexec: drop misleading information about static IP address
Including this information in commit a291e8030af493e56eddb10078705d017ab3c67b
was a mistake. Mentioned this in the original bsc#1023815.

KDUMP_NETCONFIG=eth1:static is the correct setting.

All of this information describes a workaround for some kind of a bug, which
got the netmask wrong. It's long been fixed, definitely for the current
products.

We should have fixed the product, not the documentation, in the first place.

What's worse is that the documentation is WRONG. Specifying ip= manually
without ifname= will often not get the right network interface. We're gettig
bugs from customers fiddling with the ip= settings and it's because of this bit
in the documentation.