Deployment -> Preparing the upgrade: Remove version-specific pointer sections
We cannot reliably determine if a user needs to perform a step because
we don't know the source version of the upgrade and requirements change
even throughout a SLE release, say a new Postgres or MariaDB version
(see JSC#SLE-11276).

Therefor we can only list the steps and their requirements but the
users need to figure what affects them on their own.