Modules Quick: Remove redundant support statement for modules
This overlaps with the 'Support' section below introduced in 81e31f0a20
Additionally, it includes a hardcoded link to the SLE support statement
in the release notes that we don't want to update. Unfortunately we
cannot link to the very same content in the preface of our guides as it
is commonly shared amond the guides and has no xml:id.