Upgrade guide: Disambiguate offline upgrade requiring being online
The section titled "Upgrade offline" has a curious statement about
having to be online, thus contradicting itself assuming the simplest
interpretations.

"""If the system you want to upgrade is registered with the SUSE Customer
Center, make sure to have an Internet connection during the following
procedure."""

Clarify the term "offline" in the introductory paragraph.