correction of init-scripts handling
Contrary to what was described, the init-scripts section must not be placed in the <script> part. This will cause autoyast to abort with an error message, and all validation methods will not accept this tag in the previously described location. The documentation has been corrected accordingly in both of these critical locations.