Update fuse.xml … mention archivemount-ng
There is a complete mess of various FUSE plugins dealing with archives (requirement for vifm, for example): most of them are in various stages of abandonment and support or not of fuse3. Also, not all of them (fuse-archive) are included in openSUSE. `archivemount-ng` seems to be working at least as RO file system.